告警修复

+++ net/url_request/url_request_http_job.h
@@ -360,7 +360,7 @@ class NET_EXPORT_PRIVATE URLRequestHttpJob : public URLRequestJob {
 #endif
 
 #if BUILDFLAG(ARKWEB_NETWORK_LOAD)
-  int restarted_ = 0;
+  size_t restarted_ = 0;
 #endif
 
   base::WeakPtrFactory<URLRequestHttpJob> weak_factory_{this};
